Changed doubles to floats.

+++ src/lib/libaps2.cpp
@@ -262,7 +262,7 @@ APS2_STATUS get_uptime(const char* deviceSerial, double* upTime) {
 	return aps2_getter(deviceSerial, &APS2::get_uptime, upTime);
 }
 
-APS2_STATUS get_fpga_temperature(const char* deviceSerial, double* temp) {
+APS2_STATUS get_fpga_temperature(const char* deviceSerial, float* temp) {
 	return aps2_getter(deviceSerial, &APS2::get_fpga_temperature, temp);
 }
 
@@ -356,11 +356,11 @@ APS2_STATUS get_trigger_source(const char* deviceSerial, APS2_TRIGGER_SOURCE* sr
 	return aps2_getter(deviceSerial, &APS2::get_trigger_source, src);
 }
 
-APS2_STATUS set_trigger_interval(const char* deviceSerial, double interval) {
+APS2_STATUS set_trigger_interval(const char* deviceSerial, float interval) {
 	return aps2_call(deviceSerial, &APS2::set_trigger_interval, interval);
 }
 
-APS2_STATUS get_trigger_interval(const char* deviceSerial, double* interval) {
+APS2_STATUS get_trigger_interval(const char* deviceSerial, float* interval) {
 	return aps2_getter(deviceSerial, &APS2::get_trigger_interval, interval);
 }
